SEO & Metadata Implementation for NexaFx
Added SEO metadata, Open Graph tags, Twitter cards, robots.txt, sitemap, and OG image to all public pages. Added noindex headers to dashboard and admin pages so search engines don't index user data. Set page titles for auth pages like sign-in ("Sign In") and signup ("Create Account"). Split client components that couldn't export metadata into server wrapper + client pattern. Build passes with zero errors.

Closes #243
