+++ app/page.tsx
@@ -1,34 +1,18 @@
-"use client";
+import type { Metadata } from "next";
+import LandingPageClient from "./landing-page-client";
 
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { useAuthStore } from "@/hooks/use-auth-store";
-
-import Navbar from "@/components/landing/Navbar";
-import Hero from "@/components/landing/Hero";
-import Features from "@/components/landing/Features";
-import CTA from "@/components/landing/CTA";
-import Footer from "@/components/landing/Footer";
+export const metadata: Metadata = {
+  title: "NexaFx \u2014 Multi-Currency Finance on Stellar",
+  description:
+    "Convert, deposit, and transfer currencies instantly on the Stellar blockchain.",
+  openGraph: {
+    title: "NexaFx \u2014 Multi-Currency Finance on Stellar",
+    description:
+      "Convert, deposit, and transfer currencies instantly on the Stellar blockchain.",
+    url: "https://nexafx.io",
+  },
+};
 
 export default function HomePage() {
-  const { user } = useAuthStore();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (user) router.replace("/dashboard");
-  }, [user, router]);
-
-  if (user) return null;
-
-  return (
-    <main className="min-h-screen bg-gradient-to-br from-[#A0C3FD]/10 to-[#FFE79C]/10 text-[#191c1e]">
-      <Navbar />
-      <div className="pt-24">
-        <Hero />
-        <Features />
-        <CTA />
-        <Footer />
-      </div>
-    </main>
-  );
+  return <LandingPageClient />;
 }
