+++ package.json
@@ -23,6 +23,8 @@
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "recharts": "^3.7.0",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
     "tailwind-merge": "^3.4.0",
     "zustand": "^5.0.10"
   },
