+++ public/robots.txt
@@ -1,0 +1,6 @@
+User-agent: *
+Allow: /
+Disallow: /dashboard
+Disallow: /admin
+
+Sitemap: https://nexafx.io/sitemap.xml
